Assignment5/requirements.txt
Flask==1.1.1
WTForms==2.2.1
Flask-WTF==0.14.2
google-cloud-datastore==1.9.0
cerberus==1.3.2
